refact: DTO 클래스 구조 변경 - 중첩 클래스 제거 및 request/response 패키지 분리
