+++ boss-service/src/main/java/com/happymapleday/boss/dto/request/BossPresetApplyRequest.java
@@ -1,0 +1,15 @@
+package com.happymapleday.boss.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class BossPresetApplyRequest {
+    private Long presetId;
+    private Long characterId;
+}
