+++ boss-service/src/main/java/com/happymapleday/boss/service/DesireItemService.java
@@ -1,6 +1,6 @@
 package com.happymapleday.boss.service;
 
-import com.happymapleday.boss.dto.DesireItemDto;
+import com.happymapleday.boss.dto.response.DesireItemResponse;
 import com.happymapleday.boss.repository.DesireItemRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,10 +18,10 @@ public class DesireItemService {
     private final DesireItemRepository desireItemRepository;
 
     // 특정 보스의 모든 물욕템 조회
-    public List<DesireItemDto.Response> getDesireItemsByBossId(Long bossId) {
+    public List<DesireItemResponse> getDesireItemsByBossId(Long bossId) {
         return desireItemRepository.findByBossIdOrderByItemName(bossId)
                 .stream()
-                .map(DesireItemDto.Response::fromWithRandomBoxItems)
+                .map(DesireItemResponse::fromWithRandomBoxItems)
                 .toList();
     }
 }
